pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Kara Babcock", email = "kara@tachyondecay.net"},
]
dependencies = [
    "arrow~=1.2",
    "bcrypt~=4.0",
    "beautifulsoup4~=4.11",
    "diff-match-patch~=20200713.0",
    "email-validator~=1.2",
    "flask~=2.2",
    "flask-login~=0.6",
    "flask-sqlalchemy>=3.0.2",
    "flask-wtf~=1.0",
    "gunicorn~=20.1",
    "itsdangerous~=2.1",
    "jinja2~=3.1",
    "markdown~=3.4",
    "markupsafe~=2.1",
    "passlib~=1.7",
    "python-slugify~=6.1",
    "sqlalchemy~=1.4",
    "sqlalchemy-utils~=0.38",
    "werkzeug~=2.2",
    "whoosh~=2.7",
    "wtforms~=3.0",
    "wtforms-alchemy~=0.18",
    "wtforms-components~=0.10",
    "python-dotenv~=0.20",
    "psycopg2-binary~=2.9",
    "flask-migrate~=3.1",
    "pillow~=9.2",
]
requires-python = ">=3.8"
license = {text = "N/A"}

[project.urls]
homepage = ""


[project.optional-dependencies]
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "coverage~=6.4",
    "factory-boy~=3.2",
    "pytest~=7.1",
    "pylint~=2.14",
    "pylint-pytest~=1.1",
]

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.black]
skip-string-normalization = true

[tool.coverage.run]
command_line = "-m pytest tests/"

[tool.coverage.report]
omit = [
    "__pypackages__/*",
    "tests/*",
    "venv/*",
    "/usr/*"
]

[tool.pylint]
load-plugins = "pylint_pytest"

[tool.pylint.messages_control]
disable = [
    "C0415",
    "attribute-defined-outside-init",
    "invalid-name",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "no-member",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-locals",
    "unnecessary-lambda"
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:'_app_ctx_stack':DeprecationWarning", # Remove after Flask 2.3
]